# packages/contracts/openapi-owner.json
{
...
  "info": {
    "description": "Unofficial, no SLA. Read-only river levels and discharge for the rivers flowing into the Netherlands. Rate-limited per client (429 with Retry-After). Not an official warning service. Unknown or repeated query parameters are a 400. Every 200 body carries an `attribution` array that lists exactly the sources it names. The owner API (owner view only) serves the paths of the public API with `audience: \"owner\"` in every body and `Cache-Control: private, no-store`.",
    "title": "Waterheight owner API",
    "version": "1"
  },
  "openapi": "3.1.0",
  "paths": {
    "/api/v1/beacon": {
      "post": {
        "description": "No query string. The body is at most 8,192 bytes and the content type one of the three below (parameters ignored).",
        "requestBody": {
          "content": {
            "application/csp-report": {
              "schema": {
                "additionalProperties": false,
                "description": "A legacy CSP violation report: one `csp-report` object of the standard fields, all optional.",
                "properties": {
                  "csp-report": {
                    "additionalProperties": false,
                    "properties": {
                      "blocked-uri": {
                        "maxLength": 2000,
                        "type": "string"
                      },
                      "column-number": {
                        "minimum": 0,
                        "type": "integer"
                      },
                      "disposition": {
                        "maxLength": 2000,
                        "type": "string"
                      },
                      "document-uri": {
                        "maxLength": 2000,
                        "type": "string"
                      },
                      "effective-directive": {
                        "maxLength": 2000,
                        "type": "string"
                      },
                      "line-number": {
                        "minimum": 0,
                        "type": "integer"
                      },
                      "original-policy": {
                        "maxLength": 2000,
                        "type": "string"
                      },
                      "referrer": {
                        "maxLength": 2000,
                        "type": "string"
                      },
                      "script-sample": {
                        "maxLength": 2000,
                        "type": "string"
                      },
                      "source-file": {
                        "maxLength": 2000,
                        "type": "string"
                      },
                      "status-code": {
                        "minimum": 0,
                        "type": "integer"
                      },
                      "violated-directive": {
                        "maxLength": 2000,
                        "type": "string"
                      }
                    },
                    "type": "object"
                  }
                },
                "required": [
                  "csp-report"
                ],
                "type": "object"
              }
            },
            "application/json": {
              "schema": {
                "additionalProperties": false,
                "description": "The page's own client error report.",
                "properties": {
                  "kind": {
                    "const": "client_error"
                  },
                  "message": {
                    "maxLength": 2000,
                    "type": "string"
                  },
                  "url": {
                    "maxLength": 2000,
                    "type": "string"
                  }
                },
                "required": [
                  "kind",
                  "message",
                  "url"
                ],
                "type": "object"
              }
            },
            "application/reports+json": {
              "schema": {
                "description": "A Reporting API batch: 1 to 20 reports of `type`, `age` (ms), `url`, `user_agent` and a flat `body`.",
                "items": {
                  "additionalProperties": false,
                  "properties": {
                    "age": {
                      "minimum": 0,
                      "type": "integer"
                    },
                    "body": {
                      "additionalProperties": {
                        "oneOf": [
                          {
                            "maxLength": 2000,
                            "type": "string"
                          },
                          {
                            "type": "number"
                          },
                          {
                            "type": "null"
                          }
                        ]
                      },
                      "maxProperties": 30,
                      "type": "object"
                    },
                    "type": {
                      "maxLength": 64,
                      "type": "string"
                    },
                    "url": {
                      "maxLength": 2000,
                      "type": "string"
                    },
                    "user_agent": {
                      "maxLength": 500,
                      "type": "string"
                    }
                  },
                  "required": [
                    "type",
                    "age",
                    "url",
                    "user_agent",
                    "body"
                  ],
                  "type": "object"
                },
                "maxItems": 20,
                "minItems": 1,
                "type": "array"
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": "Accepted: no body"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "A parameter is unknown, repeated, malformed or out of range, or the query is too long"
          },
          "405": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Only POST is allowed",
            "headers": {
              "Allow": {
                "description": "The methods this path takes.",
                "schema": {
                  "const": "POST",
                  "type": "string"
                }
              }
            }
          },
          "413": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "The body is longer than 8,192 bytes"
          },
          "415": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "The content type is none of the three"
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Too many requests from this client: wait for Retry-After",
            "headers": {
              "Retry-After": {
                "description": "Seconds to wait before the request is tried again (an integer, at least 1).",
                "schema": {
                  "minimum": 1,
                  "type": "integer"
                }
              }
            }
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "The data is unavailable or busy for now",
            "headers": {
              "Retry-After": {
                "description": "Seconds to wait before the request is tried again (an integer, at least 1).",
                "schema": {
                  "minimum": 1,
                  "type": "integer"
                }
              }
            }
          }
        },
        "summary": "A browser report (CSP violation, Reporting API batch or a client error): logged with control characters removed and cut, never stored"
      }
    },
    "/api/v1/frames": {
      "get": {
        "parameters": [
          {
            "description": "The start (inclusive), a whole UTC hour from ceil(displayStart). RFC 3339 with an offset (`Z` or `%2B01:00`), at most 32 characters; floored to the 10-minute UTC grid.",
            "in": "query",
            "name": "from",
            "required": true,
            "schema": {
              "maxLength": 32,
              "pattern": "^((?:19|20)[0-9]{2})-([0-9]{2})-([0-9]{2})T([0-9]{2}):([0-9]{2})(?::([0-9]{2})(?:\\.[0-9]{1,9})?)?(Z|[+-][0-9]{2}:[0-9]{2})$",
              "type": "string"
            }
          },
          {
            "description": "The end (exclusive), a whole UTC hour after from, at most 14 days after it and at most now. RFC 3339 with an offset (`Z` or `%2B01:00`), at most 32 characters; floored to the 10-minute UTC grid.",
            "in": "query",
            "name": "to",
            "required": true,
            "schema": {
              "maxLength": 32,
              "pattern": "^((?:19|20)[0-9]{2})-([0-9]{2})-([0-9]{2})T([0-9]{2}):([0-9]{2})(?::([0-9]{2})(?:\\.[0-9]{1,9})?)?(Z|[+-][0-9]{2}:[0-9]{2})$",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "step",
            "required": true,
            "schema": {
              "enum": [
                "1h"
              ],
              "type": "string"
            }
          },
          {
            "description": "The day version from `meta.dayVersions` (absent there means 1). An answer is immutable only when v is current for every day it spans and they are settled; otherwise it is cached for a short time only.",
            "in": "query",
            "name": "v",
            "required": false,
            "schema": {
              "pattern": "^[1-9][0-9]{0,5}$",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Frames"
                }
              }
            },
            "description": "Frames"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "A parameter is unknown, repeated, malformed or out of range, or the query is too long"
          },
          "405": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Only GET, HEAD is allowed",
            "headers": {
              "Allow": {
                "description": "The methods this path takes.",
                "schema": {
                  "const": "GET, HEAD",
                  "type": "string"
                }
              }
            }
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Too many requests from this client: wait for Retry-After",
            "headers": {
              "Retry-After": {
                "description": "Seconds to wait before the request is tried again (an integer, at least 1).",
                "schema": {
                  "minimum": 1,
                  "type": "integer"
                }
              }
            }
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "The data is unavailable or busy for now",
            "headers": {
              "Retry-After": {
                "description": "Seconds to wait before the request is tried again (an integer, at least 1).",
                "schema": {
                  "minimum": 1,
                  "type": "integer"
                }
              }
            }
          }
        },
        "summary": "The last hourly value of every series with data over [from, to), one row per series and one entry per hour (null where an hour has none, never carried forward); the value of bucket h is the last in [h, h + 1 h)"
      }
    },
...
    "/api/v1/meta": {
      "get": {
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Meta"
                }
              }
            },
            "description": "Meta"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "A parameter is unknown, repeated, malformed or out of range, or the query is too long"
          },
          "405": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Only GET, HEAD is allowed",
            "headers": {
              "Allow": {
                "description": "The methods this path takes.",
                "schema": {
                  "const": "GET, HEAD",
                  "type": "string"
                }
              }
            }
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Too many requests from this client: wait for Retry-After",
            "headers": {
              "Retry-After": {
                "description": "Seconds to wait before the request is tried again (an integer, at least 1).",
                "schema": {
                  "minimum": 1,
                  "type": "integer"
                }
              }
            }
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "The data is unavailable or busy for now",
            "headers": {
              "Retry-After": {
                "description": "Seconds to wait before the request is tried again (an integer, at least 1).",
                "schema": {
                  "minimum": 1,
                  "type": "integer"
                }
              }
            }
          }
        },
        "summary": "The display window, the build and the public sources with their attribution"
      }
    },
...
  }
}